requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
openai==1.10.0
google-generativeai==0.3.2
requests==2.31.0
numpy==1.26.3
nltk==3.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
tiktoken==0.5.2
textstat==0.7.3

